requirements.txt
websockets
psutil
uvicorn
websockets
numpy
soundfile
pandas
sounddevice
pyyaml

# Optional engine:
kokoro # comment out if not using Kokoro in the image
piper-tts